#475 WhatsApp: reorganize files
